+++ src/lib/services/validator.ts
@@ -1,6 +1,8 @@
 import axios from "axios";
 
-import type { ValidatorAddr } from "lib/types";
+import { CURR_THEME } from "env";
+import type { Option, ValidatorAddr, ValidatorInfo } from "lib/types";
+import { removeSpecialChars } from "lib/utils";
 
 interface ValidatorsResponse {
   validators: {
@@ -56,5 +58,31 @@ export const getValidators = async (
       },
     }),
     {}
+  );
+};
+
+export const resolveValIdentity = async (
+  apiPath: Option<string>,
+  validator: ValidatorInfo
+): Promise<string> => {
+  const githubUrl = `https://raw.githubusercontent.com/cosmostation/chainlist/master/chain/${apiPath}/moniker/${validator.validatorAddress}.png`;
+  const keybaseUrl = `https://keybase.io/_/api/1.0/user/lookup.json?key_suffix=${validator.identity}&fields=pictures`;
+  const uiAvatarsUrl = `https://ui-avatars.com/api/?name=${removeSpecialChars(
+    validator.moniker ?? ""
+  )}&background=${CURR_THEME.colors.secondary.main.replace("#", "")}&color=fff`;
+
+  return (
+    axios
+      .get(githubUrl)
+      // eslint-disable-next-line @typescript-eslint/no-unused-vars
+      .then((_) => githubUrl)
+      // eslint-disable-next-line @typescript-eslint/no-unused-vars
+      .catch(async (_) => {
+        if (validator.identity) {
+          const { data } = await axios.get(keybaseUrl);
+          return data.them[0].pictures.primary.url;
+        }
+        return uiAvatarsUrl;
+      })
   );
 };