+++ src/lib/types/validator.ts
@@ -4,4 +4,5 @@ import type { Option } from "./common";
 export interface ValidatorInfo {
   validatorAddress: ValidatorAddr;
   moniker: Option<string>;
+  identity: string;
 }